refactor: 여행 유효성 검사 커맨드로 이동 [AIR-143]
