+++ aircnc/src/main/java/com/gurudev/aircnc/domain/trip/service/command/TripCommand.java
@@ -1,5 +1,7 @@
 package com.gurudev.aircnc.domain.trip.service.command;
 
+import static com.gurudev.aircnc.exception.Preconditions.checkArgument;
+import static java.time.LocalDate.now;
 import static lombok.AccessLevel.PRIVATE;
 
 import com.gurudev.aircnc.controller.dto.TripDto.TripReserveRequest;
@@ -15,6 +17,8 @@ public final class TripCommand {
   @Getter
   public static class TripReserveCommand {
 
+    public static final int TRIP_TOTAL_PRICE_MIN_VALUE = 10000;
+
     private Long guestId;
     private Long roomId;
     private LocalDate checkIn;
@@ -24,6 +28,15 @@ public final class TripCommand {
 
     public TripReserveCommand(Long guestId, Long roomId, LocalDate checkIn,
         LocalDate checkOut, int headCount, int totalPrice) {
+      checkArgument(checkOut.isAfter(checkIn), "체크아웃은 체크인 이전이 될 수 없습니다");
+      checkArgument(checkIn.isEqual(now()) || checkIn.isAfter(now()),
+          "체크인 날짜는" + now() + " 이전이 될 수 없습니다.");
+
+      checkArgument(totalPrice >= TRIP_TOTAL_PRICE_MIN_VALUE,
+          String.format("총 가격은 %d원 미만이 될 수 없습니다", TRIP_TOTAL_PRICE_MIN_VALUE));
+
+      checkArgument(headCount >= 1, "인원은 1명 이상이여야 합니다");
+
       this.guestId = guestId;
       this.roomId = roomId;
       this.checkIn = checkIn;
